Rework discount UI
