+++ packages/admin/resources/views/livewire/components/discounts/show.blade.php
@@ -13,49 +13,7 @@
         </h1>
     </div>
 
-    <div class="pb-24 mt-8 lg:gap-8 lg:flex lg:items-start">
-        <div class="space-y-6 lg:flex-1">
-            <div class="space-y-6">
-                <div id="basic-information">
-                    @include('adminhub::partials.forms.discount.basic-information')
-                </div>
-
-                <div id="availability">
-                    @include('adminhub::partials.availability', [
-                        'channels' => true,
-                        'type' => 'discount',
-                        'customerGroups' => true,
-                        'customerGroupType' => 'toggle',
-                    ])
-                </div>
-            </div>
-        </div>
-        <x-hub::layout.page-menu>
-            @foreach ($this->sideMenu as $item)
-                <a href="#{{ $item['id'] }}"
-                   @class([
-                       'flex items-center gap-2 p-2 rounded text-gray-500',
-                       'hover:bg-blue-50 hover:text-blue-700' => empty($item['has_errors']),
-                       'text-red-600 bg-red-50' => !empty($item['has_errors']),
-                   ])
-                   aria-current="page"
-                   x-data="{ linkId: '#{{ $item['id'] }}' }"
-                   :class="{
-                       'bg-blue-50 text-blue-700 hover:text-blue-600': linkId === activeAnchorLink
-                   }"
-                   x-on:click="activeAnchorLink = linkId">
-                    @if (!empty($item['has_errors']))
-                        <x-hub::icon ref="exclamation-circle"
-                                     class="w-4 text-red-600" />
-                    @endif
-
-                    <span class="text-sm font-medium">
-                        {{ $item['title'] }}
-                    </span>
-                </a>
-            @endforeach
-        </x-hub::layout.page-menu>
-    </div>
+    @include('adminhub::partials.forms.discount')
 
     {{-- <div class="space-y-4">
