+++ packages/admin/src/AdminHubServiceProvider.php
@@ -11,6 +11,7 @@ use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 use Livewire\Livewire;
+use Lunar\DiscountTypes\Discount;
 use Lunar\Hub\Auth\Manifest;
 use Lunar\Hub\Base\ActivityLog\Manifest as ActivityLogManifest;
 use Lunar\Hub\Console\Commands\InstallHub;
@@ -37,7 +38,9 @@ use Lunar\Hub\Http\Livewire\Components\Discounts\DiscountCreate;
 use Lunar\Hub\Http\Livewire\Components\Discounts\DiscountShow;
 use Lunar\Hub\Http\Livewire\Components\Discounts\DiscountsIndex;
 use Lunar\Hub\Http\Livewire\Components\Discounts\DiscountsTable;
+use Lunar\Hub\Http\Livewire\Components\Discounts\Types\BuyXGetY;
 use Lunar\Hub\Http\Livewire\Components\Discounts\Types\Coupon;
+use Lunar\Hub\Http\Livewire\Components\Discounts\Types\Discount as TypesDiscount;
 use Lunar\Hub\Http\Livewire\Components\Discounts\Types\ProductDiscount;
 use Lunar\Hub\Http\Livewire\Components\Orders\EmailNotification;
 use Lunar\Hub\Http\Livewire\Components\Orders\OrderCapture;
@@ -445,8 +448,8 @@ class AdminHubServiceProvider extends ServiceProvider
         Livewire::component('hub.components.discounts.create', DiscountCreate::class);
         Livewire::component('hub.components.discounts.table', DiscountsTable::class);
 
-        Livewire::component('lunar.hub.http.livewire.components.discounts.types.coupon', Coupon::class);
-        Livewire::component('lunar.hub.http.livewire.components.discounts.types.product-discount', ProductDiscount::class);
+        Livewire::component('lunar.hub.http.livewire.components.discounts.types.discount', TypesDiscount::class);
+        Livewire::component('lunar.hub.http.livewire.components.discounts.types.buy-x-get-y', BuyXGetY::class);
     }
 
     /**
