+++ packages/admin/src/Http/Livewire/Components/Discounts/DiscountCreate.php
@@ -3,6 +3,7 @@
 namespace Lunar\Hub\Http\Livewire\Components\Discounts;
 
 use Illuminate\Support\Str;
+use Lunar\DiscountTypes\Discount as DiscountTypesDiscount;
 use Lunar\DiscountTypes\ProductDiscount;
 use Lunar\Models\Currency;
 use Lunar\Models\Discount;
@@ -22,12 +23,13 @@ class DiscountCreate extends AbstractDiscount
     public function mount()
     {
         $this->discount = new Discount([
-            'type' => ProductDiscount::class,
+            'type' => DiscountTypesDiscount::class,
             'starts_at' => now()->startOfHour(),
             'data' => [],
         ]);
 
         $this->currency = Currency::getDefault();
+            $this->syncAvailability();
     }
 
     /**
